+++ packages/awino-ui/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add stats component to landing page
 - Add info component to landing page
 - Add benefit component to landing page
+- Add guide component to landing page
 
 ## 03.03.2022